Test unbound session scenarios
