+++ pom.xml
@@ -41,7 +41,7 @@
 
     <properties>
         <revision>999.0.0-SNAPSHOT</revision>
-        <keycloak.version>26.5.4</keycloak.version>
+        <keycloak.version>26.5.6</keycloak.version>
         <maven.compiler.release>21</maven.compiler.release>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
