+++ src/main/resources/theme/keycloak.v2+oid4vp/login/messages/messages_en.properties
@@ -1,2 +1,4 @@
 oid4vpLoginBtnMessage=Sign in with a wallet
 oid4vpScanPageTitle=Scan the QR code to sign in to your account
+oid4vpSameDevicePrompt=Have your wallet on this device?
+oid4vpSameDeviceLinkLbl=Click to sign in with your wallet
