+++ src/test/java/io/github/adorsysgis/keycloak/protocol/oid4vc/oid4vp/OID4VPUserAuthEndpointTest.java
@@ -6,71 +6,44 @@ import static io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.authenticator
 import static io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.authenticator.SdJwtCredentialConstrainer.QueryMap;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.authenticator.SdJwtCredentialConstrainerTest;
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.RequestObject;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.ResponseObject;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.dcql.Credential;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.dcql.DcqlQuery;
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.dto.AuthorizationContext;
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.dto.AuthorizationContextStatus;
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.dto.ProcessingError;
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.prex.Descriptor;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.prex.InputDescriptor;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.prex.PresentationDefinition;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.model.prex.PresentationSubmission;
-import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.service.AuthorizationResponseService;
 import io.github.adorsysgis.keycloak.protocol.oid4vc.oid4vp.utils.SdJwtVPTestUtils;
 import java.io.ByteArrayInputStream;
-import java.io.IOException;
 import java.net.URI;
-import java.nio.charset.StandardCharsets;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
-import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import org.apache.http.HttpHeaders;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
-import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.utils.URIBuilder;
-import org.apache.http.impl.client.BasicCookieStore;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.message.BasicNameValuePair;
-import org.jboss.resteasy.specimpl.ResteasyUriInfo;
 import org.junit.jupiter.api.Test;
 import org.keycloak.OAuth2Constants;
 import org.keycloak.OAuthErrorException;
-import org.keycloak.TokenVerifier;
-import org.keycloak.common.VerificationException;
-import org.keycloak.jose.jwk.JWK;
 import org.keycloak.jose.jws.JWSHeader;
 import org.keycloak.jose.jws.JWSInput;
 import org.keycloak.representations.AccessToken;
 import org.keycloak.representations.JsonWebToken;
 import org.keycloak.representations.idm.OAuth2ErrorRepresentation;
-import org.keycloak.util.JsonSerialization;
 
 /**
  * Testing OpenID4VP user authentication via presentation of SD-JWT identity credentials.
  *
  * @author <a href="mailto:Ingrid.Kamga@adorsys.com">Ingrid Kamga</a>
  */
-public class OID4VPUserAuthEndpointTest extends OID4VPBaseKeycloakTest {
+public class OID4VPUserAuthEndpointTest extends OID4VPBaseUserAuthEndpointTest {
 
     public static final String VCT_CONFIG_ALT = "https://example.com/vct-alt";
-
-    private final SdJwtVPTestUtils sdJwtVPTestUtils = new SdJwtVPTestUtils(keycloak, getActiveTestRealm());
 
     @Test
     public void shouldProduceAuthorizationRequests() throws Exception {
@@ -566,384 +539,6 @@ public class OID4VPUserAuthEndpointTest extends OID4VPBaseKeycloakTest {
         for (String invalidAud : invalidAuds) {
             testFailAuthentication_InvalidKbJwt(
                     null, invalidAud, null, null, "claim 'aud' does not match actual value");
-        }
-    }
-
-    /**
-     * Helper for successful flows.
-     */
-    private String testSuccessfulAuthentication(String sdJwt, TestOpts opts) throws Exception {
-        // Retrieve an authorization request
-        AuthorizationContext authContext = requestAuthorizationRequest();
-        RequestObject requestObject = resolveRequestObject(authContext.getAuthorizationRequest());
-
-        // Prepare and send the OpenID4VP response to Keycloak
-        HttpResponse response = sendAuthorizationResponse(sdJwt, requestObject, opts);
-        assertEquals(HttpStatus.SC_OK, response.getStatusLine().getStatusCode());
-
-        // Check auth status
-        HttpResponse statusResponse = fetchAuthenticationStatus(authContext.getTransactionId());
-        AuthorizationContext statusPayload = parseAuthorizationContext(statusResponse);
-        assertEquals(AuthorizationContextStatus.SUCCESS, statusPayload.getStatus());
-
-        // Exchange authorization code for access token
-        String authCode = statusPayload.getAuthorizationCode();
-        assertNotNull(authCode, "Authorization code should not be null");
-        if (opts.shouldRetrieveAccessToken()) {
-            assertAuthenticatingUser(opts, authCode);
-        }
-
-        // Bubble up authorization code
-        return authCode;
-    }
-
-    private void assertAuthenticatingUser(TestOpts opts, String authCode) throws VerificationException, IOException {
-        String accessTokenStr = requestAccessToken(authCode, opts.shouldEnforceRedirectUri());
-        AccessToken accessToken =
-                TokenVerifier.create(accessTokenStr, AccessToken.class).getToken();
-
-        // Assert authenticating user
-        assertEquals(opts.getTestUser(), accessToken.getPreferredUsername());
-
-        // Assert token issuer
-        assertEquals(getTestRealmEndpoint(), accessToken.getIssuer());
-    }
-
-    /**
-     * Helper for failing flows.
-     */
-    private void testFailingAuthentication(
-            String sdJwt, TestOpts opts, int httpStatus, String expectedError, String expectedErrorDescription)
-            throws Exception {
-        // Retrieve an authorization request
-        AuthorizationContext authContext = requestAuthorizationRequest();
-        RequestObject requestObject = resolveRequestObject(authContext.getAuthorizationRequest());
-        if (opts.getOverridePresentationDefinitionId() != null) {
-            requestObject.getPresentationDefinition().setId(opts.getOverridePresentationDefinitionId());
-        }
-
-        // Prepare and send the OpenID4VP response to Keycloak
-        HttpResponse response = sendAuthorizationResponse(sdJwt, requestObject, opts);
-
-        // Run assertions
-        assertFailingAuthentication(
-                response, authContext.getTransactionId(), httpStatus, expectedError, expectedErrorDescription);
-    }
-
-    /**
-     * Helper for failing flows (from VP token).
-     */
-    private void testFailingAuthenticationWithVPToken(
-            String sdJwtVpToken,
-            RequestObject requestObject,
-            String transactionId,
-            int httpStatus,
-            String expectedError,
-            String expectedErrorDescription)
-            throws Exception {
-        // Prepare and send the OpenID4VP response to Keycloak
-        HttpResponse response = sendAuthorizationResponseWithVPToken(sdJwtVpToken, requestObject, new TestOpts());
-
-        // Run assertions
-        assertFailingAuthentication(response, transactionId, httpStatus, expectedError, expectedErrorDescription);
-    }
-
-    /**
-     * Helper for asserting failing flows.
-     */
-    private void assertFailingAuthentication(
-            HttpResponse postAuthResponse,
-            String transactionId,
-            int httpStatus,
-            String expectedError,
-            String expectedErrorDescription)
-            throws Exception {
-        assertEquals(httpStatus, postAuthResponse.getStatusLine().getStatusCode());
-        OAuth2ErrorRepresentation errorRep = parseErrorResponse(postAuthResponse);
-        assertEquals(expectedError, errorRep.getError());
-        assertTrue(errorRep.getErrorDescription().contains(expectedErrorDescription));
-
-        // Check and assert auth status
-        HttpResponse statusResponse = fetchAuthenticationStatus(transactionId);
-        AuthorizationContext statusPayload = parseAuthorizationContext(statusResponse);
-        assertEquals(AuthorizationContextStatus.ERROR, statusPayload.getStatus());
-        assertEquals(expectedError, statusPayload.getError().getErrorString());
-        assertTrue(statusPayload.getErrorDescription().contains(expectedErrorDescription));
-    }
-
-    /**
-     * Helper for failing flows (Invalid KB-JWTs).
-     */
-    private void testFailAuthentication_InvalidKbJwt(
-            String overrideNonce, String overrideAud, JWK holderkey, Integer kbJwtLifespanSecs, String errorMessage)
-            throws Exception {
-        // Request a SD-JWT credential from Keycloak to use for authentication
-        String sdJwt = sdJwtVPTestUtils.requestSdJwtCredential(VCT_CONFIG_DEFAULT, TEST_USER);
-
-        // Retrieve an authorization request
-        AuthorizationContext authContext = requestAuthorizationRequest();
-        RequestObject requestObject = resolveRequestObject(authContext.getAuthorizationRequest());
-
-        // Prepare SD-JWT VP tokens with invalid KB-JWTs
-        String sdJwtVpToken = sdJwtVPTestUtils.presentSdJwt(
-                sdJwt,
-                overrideNonce == null ? requestObject.getNonce() : overrideNonce,
-                overrideAud == null ? requestObject.getClientId() : overrideAud,
-                holderkey == null ? SdJwtVPTestUtils.getUserJwk() : holderkey,
-                kbJwtLifespanSecs == null ? SdJwtVPTestUtils.KB_JWT_LIFESPAN_SECS : kbJwtLifespanSecs);
-
-        // Proceed to authentication
-        testFailingAuthenticationWithVPToken(
-                sdJwtVpToken,
-                requestObject,
-                authContext.getTransactionId(),
-                HttpStatus.SC_UNAUTHORIZED,
-                ProcessingError.VP_TOKEN_AUTH_ERROR.getErrorString(),
-                errorMessage);
-    }
-
-    /**
-     * Sends an OpenID4VP response to Keycloak, producing an SD-JWT verifiable presentation.
-     */
-    private HttpResponse sendAuthorizationResponse(String sdJwt, RequestObject requestObject, TestOpts opts)
-            throws Exception {
-        // Prepare a valid SD-JWT verifiable presentation
-        String sdJwtVpToken = sdJwtVPTestUtils.presentSdJwt(
-                sdJwt,
-                requestObject.getNonce(),
-                opts.getOverridePresentationAud() == null
-                        ? requestObject.getClientId()
-                        : opts.getOverridePresentationAud(),
-                SdJwtVPTestUtils.getUserJwk());
-
-        // Base64-encode the SD-JWT VP token if requested
-        if (opts.getShouldBase64EncodeVpToken()) {
-            byte[] bytes = sdJwtVpToken.getBytes(StandardCharsets.UTF_8);
-            sdJwtVpToken = Base64.getUrlEncoder().encodeToString(bytes);
-        }
-
-        // Send the OpenID4VP response to Keycloak
-        return sendAuthorizationResponseWithVPToken(sdJwtVpToken, requestObject, opts);
-    }
-
-    /**
-     * Sends an OpenID4VP response to Keycloak, producing an SD-JWT verifiable presentation.
-     */
-    private HttpResponse sendAuthorizationResponseWithVPToken(
-            String sdJwtVpToken, RequestObject requestObject, TestOpts opts) throws Exception {
-        // Wrap the SD-JWT VP in an OpenID4VP response
-        List<BasicNameValuePair> oid4vpResponse;
-        if (opts.shouldPrepareLegacyResponse()) {
-            oid4vpResponse = prepareLegacyOpenID4VPResponse(sdJwtVpToken, requestObject, opts);
-        } else {
-            oid4vpResponse = prepareOpenID4VPResponse(sdJwtVpToken, requestObject);
-        }
-
-        // Send the OpenID4VP response to Keycloak
-        String url = getOid4vpEndpoint("/response");
-        HttpPost httpPost = new HttpPost(url);
-        httpPost.setEntity(new UrlEncodedFormEntity(oid4vpResponse));
-        return httpClient.execute(httpPost);
-    }
-
-    /**
-     * Prepare the OpenID4VP response object to be sent to Keycloak.
-     *
-     * @param sdJwtVpToken  the SD-JWT verifiable presentation token
-     * @param requestObject the request object containing the presentation definition
-     */
-    private List<BasicNameValuePair> prepareOpenID4VPResponse(String sdJwtVpToken, RequestObject requestObject)
-            throws IOException {
-        // Build presentation submission
-        DcqlQuery dcqlQuery = requestObject.getDcqlQuery();
-        Credential credentialQuery = dcqlQuery.getCredentials().getFirst();
-        var vpTokenMap = Map.of(credentialQuery.getId(), List.of(sdJwtVpToken));
-
-        // Compose the response object as form-urlencoded parameters
-        return new ArrayList<>(List.of(
-                new BasicNameValuePair(ResponseObject.VP_TOKEN_KEY, JsonSerialization.writeValueAsString(vpTokenMap)),
-                new BasicNameValuePair(ResponseObject.STATE_KEY, requestObject.getState())));
-    }
-
-    /**
-     * Prepare the OpenID4VP response object to be sent to Keycloak (Legacy).
-     *
-     * @param sdJwtVpToken  the SD-JWT verifiable presentation token
-     * @param requestObject the request object containing the presentation definition
-     */
-    private List<BasicNameValuePair> prepareLegacyOpenID4VPResponse(
-            String sdJwtVpToken, RequestObject requestObject, TestOpts opts) throws IOException {
-        // Build presentation submission
-
-        PresentationDefinition definition = requestObject.getPresentationDefinition();
-        InputDescriptor inputDescriptor = definition.getInputDescriptors().getFirst();
-
-        PresentationSubmission submission = new PresentationSubmission();
-        submission.setId(UUID.randomUUID().toString());
-        submission.setDefinitionId(definition.getId());
-
-        Descriptor descriptor = new Descriptor();
-        descriptor.setId(inputDescriptor.getId());
-        descriptor.setFormat(
-                opts.getOverrideDescriptorFormat() == null
-                        ? Descriptor.Format.VC_SD_JWT
-                        : opts.getOverrideDescriptorFormat());
-        descriptor.setPath(
-                opts.getOverrideDescriptorPath() == null
-                        ? AuthorizationResponseService.JSON_PATH_ROOT
-                        : opts.getOverrideDescriptorPath());
-        submission.setDescriptorMap(List.of(descriptor));
-
-        // Compose the response object as form-urlencoded parameters
-
-        return new ArrayList<>(List.of(
-                new BasicNameValuePair(ResponseObject.VP_TOKEN_KEY, sdJwtVpToken),
-                new BasicNameValuePair(
-                        ResponseObject.PRESENTATION_SUBMISSION_KEY, JsonSerialization.writeValueAsString(submission)),
-                new BasicNameValuePair(ResponseObject.STATE_KEY, requestObject.getState())));
-    }
-
-    @Test
-    public void shouldAuthenticateSuccessfully_InOIDCFlow() throws Exception {
-        // Request a valid SD-JWT credential from Keycloak to use for authentication
-        String sdJwt = sdJwtVPTestUtils.requestSdJwtCredential(VCT_CONFIG_DEFAULT, TEST_USER);
-
-        // Collect OIDC session data
-        FormData formData = getFreshOid4vpFormActionUrl();
-        String actionURI = formData.actionUrl();
-        BasicCookieStore cookieStore = formData.cookieStore();
-
-        // Proceed to authentication
-        TestOpts opts = TestOpts.getDefault().setShouldRetrieveAccessToken(false);
-        String authCode = testSuccessfulAuthentication(sdJwt, opts);
-        BasicNameValuePair codeParam = new BasicNameValuePair(OAuth2Constants.CODE, authCode);
-
-        // Continue OIDC flow with auth code
-        try (CloseableHttpClient httpClient =
-                HttpClientBuilder.create().setDefaultCookieStore(cookieStore).build()) {
-            HttpPost httpPost = new HttpPost(actionURI);
-            httpPost.setEntity(new UrlEncodedFormEntity(List.of(codeParam)));
-            HttpResponse httpResponse = httpClient.execute(httpPost);
-            assertEquals(
-                    HttpStatus.SC_MOVED_TEMPORARILY,
-                    httpResponse.getStatusLine().getStatusCode());
-
-            String redirectUri =
-                    httpResponse.getFirstHeader(HttpHeaders.LOCATION).getValue();
-            assertTrue(redirectUri.startsWith(TEST_CLIENT_REDIRECT_URI));
-
-            // Extract the authorization code from the redirect URI
-            ResteasyUriInfo uriInfo = new ResteasyUriInfo(URI.create(redirectUri));
-            String freshAuthCode = uriInfo.getQueryParameters().getFirst(OAuth2Constants.CODE);
-
-            // Assert the validity of the fresh auth code
-            assertAuthenticatingUser(opts.setShouldEnforceRedirectUri(true), freshAuthCode);
-            assertNotEquals("New code must be issued", authCode, freshAuthCode);
-
-            // TODO: A login method param must be appended to the redirect URI
-            // String loginMethod = uriInfo.getQueryParameters().getFirst(PARAM_LOGIN_METHOD);
-            // assertEquals(OID4VP_AUTH_LOGIN_PATH, loginMethod);
-        }
-    }
-
-    /**
-     * POJO for test options.
-     */
-    static class TestOpts {
-
-        private String testUser = TEST_USER;
-        private boolean shouldBase64EncodeVpToken;
-        private boolean shouldRetrieveAccessToken = true;
-        private boolean shouldPrepareLegacyResponse = true;
-        private boolean shouldEnforceRedirectUri = false;
-        private String overridePresentationDefinitionId;
-        private String overridePresentationAud;
-        private Descriptor.Format overrideDescriptorFormat;
-        private String overrideDescriptorPath;
-
-        public static TestOpts getDefault() {
-            return new TestOpts();
-        }
-
-        public String getTestUser() {
-            return testUser;
-        }
-
-        public TestOpts setTestUser(String testUser) {
-            this.testUser = testUser;
-            return this;
-        }
-
-        public boolean getShouldBase64EncodeVpToken() {
-            return shouldBase64EncodeVpToken;
-        }
-
-        public TestOpts setShouldBase64EncodeVpToken(boolean shouldBase64EncodeVpToken) {
-            this.shouldBase64EncodeVpToken = shouldBase64EncodeVpToken;
-            return this;
-        }
-
-        public boolean shouldRetrieveAccessToken() {
-            return shouldRetrieveAccessToken;
-        }
-
-        public TestOpts setShouldRetrieveAccessToken(boolean retrieveAccessToken) {
-            this.shouldRetrieveAccessToken = retrieveAccessToken;
-            return this;
-        }
-
-        public boolean shouldPrepareLegacyResponse() {
-            return shouldPrepareLegacyResponse;
-        }
-
-        public TestOpts setShouldPrepareLegacyResponse(boolean shouldPrepareLegacyResponse) {
-            this.shouldPrepareLegacyResponse = shouldPrepareLegacyResponse;
-            return this;
-        }
-
-        public boolean shouldEnforceRedirectUri() {
-            return shouldEnforceRedirectUri;
-        }
-
-        public TestOpts setShouldEnforceRedirectUri(boolean shouldEnforceRedirectUri) {
-            this.shouldEnforceRedirectUri = shouldEnforceRedirectUri;
-            return this;
-        }
-
-        public String getOverridePresentationDefinitionId() {
-            return overridePresentationDefinitionId;
-        }
-
-        public TestOpts setOverridePresentationDefinitionId(String overridePresentationDefinitionId) {
-            this.overridePresentationDefinitionId = overridePresentationDefinitionId;
-            return this;
-        }
-
-        public String getOverridePresentationAud() {
-            return overridePresentationAud;
-        }
-
-        public TestOpts setOverridePresentationAud(String overridePresentationAud) {
-            this.overridePresentationAud = overridePresentationAud;
-            return this;
-        }
-
-        public Descriptor.Format getOverrideDescriptorFormat() {
-            return overrideDescriptorFormat;
-        }
-
-        public TestOpts setOverrideDescriptorFormat(Descriptor.Format overrideDescriptorFormat) {
-            this.overrideDescriptorFormat = overrideDescriptorFormat;
-            return this;
-        }
-
-        public String getOverrideDescriptorPath() {
-            return overrideDescriptorPath;
-        }
-
-        public TestOpts setOverrideDescriptorPath(String overrideDescriptorPath) {
-            this.overrideDescriptorPath = overrideDescriptorPath;
-            return this;
         }
     }
 }
